TTP-450 Revise C-CDA Validator UI
Validator Home
Please change the text on the landing page to the following:
This area provides tools for testing conformance of artifacts to industry standards and specific criteria. There are both context-free validators for general testing and validators for the criteria of the 2015 Edition.

C-CDA R1.1 Validator
Please change the title label to: "C-CDA R1.1 Validator" (remove the ONC MDHT Tool)
Please change the default drop-down label to "Select validation criteria…"
Please put the "Non-specific CCDA" at the top of the listing if possible.

C-CDA R2.1 Validator

See if you can slide the view down to the validation report on completion of the validation run (when finished change the view to the report as it is off the screen)

iFrame label should be "R2.1 Validator"

Change the test on the context free validation to "To validate with the C-CDA Implementation Guide attach or drag your C-CDA into the tool and press Validate."

Eliminate the use of the term "S&CC" as follows: "Select the ONC 2015 Edition criteria:"

Please change the default drop-down label to "Select scenario file…"